requirements.txt
Flask
gunicorn
python-binance
supabase
